+++ TODO
@@ -1,4 +1,3 @@
 - Finish API documentation
-- Add multiple RPC calls in client.c
 - Make rpcclnt_mainloop pthread_create friendly
 - Make rpcclnt_destroy 'thread-safe'
