+++ tests/client.c
@@ -1,5 +1,4 @@
 #include "calc.pb-c.h"
-#include "../pbrpc.h"
 #include "../pbrpc-clnt.h"
 
 #include <inttypes.h>
